Delete event handling refactor
