+++ src/Equinor.ProCoSys.Preservation.Command/Events/JourneyDeleteEvent.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using Equinor.ProCoSys.Preservation.Command.Events.EntityNames;
+
+namespace Equinor.ProCoSys.Preservation.Command.Events;
+
+[JourneyEntityName]
+public class JourneyDeleteEvent : DeleteEvent
+{
+    public JourneyDeleteEvent(Guid guid, string plant) : base(guid, plant, null) {}
+}
