+++ src/Equinor.ProCoSys.Preservation.Command/Events/ModeDeleteEvent.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using Equinor.ProCoSys.Preservation.Command.Events.EntityNames;
+
+namespace Equinor.ProCoSys.Preservation.Command.Events;
+
+[ModeEntityName]
+public class ModeDeleteEvent : DeleteEvent
+{
+    public ModeDeleteEvent(Guid guid, string plant) : base(guid, plant, null) {}
+}
