+++ src/Equinor.ProCoSys.Preservation.Command/Events/PreservationPeriodDeleteEvent.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using Equinor.ProCoSys.Preservation.Command.Events.EntityNames;
+
+namespace Equinor.ProCoSys.Preservation.Command.Events;
+
+[PreservationPeriodEntityName]
+public class PreservationPeriodDeleteEvent : DeleteEvent
+{
+    public PreservationPeriodDeleteEvent(Guid guid, string plant) : base(guid, plant, null) {}
+}
