+++ src/Equinor.ProCoSys.Preservation.Command/Events/RequirementTypeDeleteEvent.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using Equinor.ProCoSys.Preservation.Command.Events.EntityNames;
+
+namespace Equinor.ProCoSys.Preservation.Command.Events;
+
+[RequirementTypeEntityName]
+public class RequirementTypeDeleteEvent : DeleteEvent
+{
+    public RequirementTypeDeleteEvent(Guid guid, string plant) : base(guid, plant, null) {}
+}
