+++ src/Equinor.ProCoSys.Preservation.Command/Events/TagDeleteEvent.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using Equinor.ProCoSys.Preservation.Command.Events.EntityNames;
+
+namespace Equinor.ProCoSys.Preservation.Command.Events;
+
+[TagEntityName]
+public class TagDeleteEvent : DeleteEvent
+{
+    public TagDeleteEvent(Guid guid, string plant, string projectName) : base(guid, plant, projectName)
+    {
+        
+    }
+}
