+++ src/Equinor.ProCoSys.Preservation.Domain/Events/IPlantEntityEvent.cs
@@ -2,7 +2,7 @@
 
 namespace Equinor.ProCoSys.Preservation.Domain.Events;
 
-public interface IPlantEntityEvent<out T> where T : PlantEntityBase
+public interface IPlantEntityEvent<out T> where T : PlantEntityBase, IHaveGuid
 {
     T Entity { get; }
 }
